% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{The name of the file that the function will read into R.}
}
\value{
This function removes events in the BCE era (49 out of the 6416 events) and returns a cleaned NOAA data
frame, with the following additional changes:
\enumerate{
\item A Date column created by uniting the year, month, and day variables, and converting it to the Date class
\item Latitude and Longitude variables converted to numeric class
}
}
\description{
This function further cleans the NOAA data frame produced by the eq_location_clean function.
}
\note{
Example: eq_clean_data(earthquakes_data_clean1)

If the file  does not exist (e.g. earthquakes), attempting to read this file
will generate an error.
try(eq_clean_data(earthquakes))

Attempting to read a non-tibble file will also generate an error.
try(eq_clean_data("earthquakes_data_clean1.xlsx"))
}
