% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The input database.}
}
\value{
This function constructs a character string for each earthquake that includes location, magnitude, and total
number of deaths (all boldface labels). If an earthquake is missing values for any of these, both label and value
will be skipped for that element of tag.
}
\description{
This function creates an HTML label that can be used as annotation text in interactive maps of NOAA earthquake epicenters.
}
\note{
Example: library(magrittr)
dplyr::filter(earthquakes_data_clean2, Country == "MEXICO" & lubridate::year(Date) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col="popup_text", -120, -60, 0, 40)

If the selected annotation columns do not exist, attempting to use these columns will
generate an error.
try(eq_create_label(annot_col = "Columns"))
}
