% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\name{eq_filter_data}
\alias{eq_filter_data}
\title{eq_filter_data}
\usage{
eq_filter_data(filename, xmin, xmax, columns)
}
\arguments{
\item{filename}{The name of the file that the function will read into R.}

\item{xmin}{Minimum date for filtering}

\item{xmax}{Maximum date for filtering}

\item{columns}{Columns of interest to select}
}
\value{
This function returns the final NOAA data frame for analysis, with the following additional changes:
\enumerate{
\item Filtering a subset of NOAA data between a minimum and maximum date
\item Additional of columns of interest
\item Selection of all columns of interest
\item Filtering of NOAA data for complete cases
}
}
\description{
This function filters the NOAA data frame produced by the eq_clean_data function.
}
\note{
Example: eq_filter_data(earthquakes_data_clean2, xmin = "1950-01-01", xmax = "1955-01-01",
c("Date", "xend", "y", "yend", "Mag", "Deaths", "Country"))

If the file  does not exist (e.g. earthquakes), attempting to read this file
will generate an error.
try(eq_filter_data(earthquakes, xmin = "1950-01-01", xmax = "1955-01-01",
c("Date", "xend", "y", "yend", "Mag", "Deaths", "Country")))
}
