% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(filename)
}
\arguments{
\item{filename}{The name of the file that the function will read into R.}
}
\value{
This function strips out the country names (including the colon) from the `Location Name' variable
and converts the resulting string(s) to title case (as opposed to all caps).
}
\description{
This function cleans the LOCATION_NAME column in the raw NOAA data frame.
}
\note{
Example: eq_location_clean("earthquakes_-2150to2024.csv")

If the file  does not exist (e.g. earthquakes_-2150to2050.csv), attempting to read this file
will generate an error.
try(eq_location_clean("earthquakes_-2150to2024.csv"))

Attempting to read a non-CSV file will also generate an error.
try(eq_location_clean("earthquakes_-2150to2024.xlsx"))
}
