% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(eq_data, annot_col, x1, x2, y1, y2)
}
\arguments{
\item{eq_data}{The name of the file that the function will read into R.}

\item{annot_col}{The column to use for the annotation in the pop-up.}

\item{x1}{Western longitude limit for plot by 5 degree increments.}

\item{x2}{Eastern longitude limit for plot by 5 degree increments.}

\item{y1}{Southern latitude limit for plot by 5 degree increments.}

\item{y2}{Northern latitude limit for plot by 5 degree increments.}
}
\value{
This function maps earthquake epicenters (LATITUDE/LONGITUDE) and annotates each point within a pop up window
containing annotation data stored in a column of the data frame. The user can also choose which column is used for
the annotation in the pop-up.
}
\description{
This function uses NOAA data to create interactive maps of select NOAA earthquake epicenters.
}
\note{
Example: library(magrittr)
dplyr::filter(earthquakes_data_clean2, Country == "MEXICO" & lubridate::year(Date) >= 2000) \%>\%
eq_map(annot_col="Date", -120, -60, 0, 40)

If the selected annotation column  does not exist, attempting to use this column will
generate an error.
try(eq_map(annot_col = "Column",0,180,30,90))
}
