% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The mapping of variables}

\item{data}{The input database}

\item{stat}{The stat to use}

\item{position}{The position to use}

\item{...}{Additional parameters}

\item{na.rm}{Whether to remove NA values}

\item{show.legend}{Whether to show the legend}

\item{inherit.aes}{Whether to inherit aesthetics}
}
\value{
This function returns a layer, that consists of the time line of selected earthquakes ranging from
min to max dates (with a point for each earthquake), using GeomTimeline as default geom class.
}
\description{
This function uses GeomTimeline as the default geom class and plots a time line of selected earthquakes.
}
\note{
Example: ggplot() + geom_timeline(data=earthquakes_data_final, aes(Date, y=0.2, size=Mag, colour=Deaths), alpha=0.3) +
theme(axis.text.x=element_text(angle=60, hjust=1))

If the file  does not exist (e.g. earthquakes), attempting to read this file will generate an error.
try(ggplot() + geom_timeline(data=earthquakes, aes(Date, y=0.2, size=Mag, colour=Deaths), alpha=0.3) +
theme(axis.text.x=element_text(angle=60, hjust=1)))
}
