% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The mapping of variables}

\item{data}{The input database}

\item{stat}{The stat to use}

\item{position}{The position to use}

\item{...}{Additional parameters}

\item{na.rm}{Whether to remove NA values}

\item{show.legend}{Whether to show the legend}

\item{inherit.aes}{Whether to inherit aesthetics}
}
\value{
This function returns a layer, that consists of the time line of selected earthquakes ranging from
min to max dates (with a point for each earthquake), and a vertical line to selected data points with a
text annotation, using GeomTimeline as default geom class.
}
\description{
This function uses GeomTimelineLabel as the default geom class and plots a time line of selected earthquakes.
}
\note{
Example: gplot() + geom_timeline_label(data=earthquakes_data_final, aes(Date, y=0.2, xend=xend, yend=yend, size=Mag, colour=Deaths,
label=Country), alpha=0.3) +
annotate("text", x=earthquakes_data_final$Date, y=0.42, label=earthquakes_data_final$Country, vjust=0, hjust=-0, size=2, angle=45) +
ylim(c(0,1)) + geom_hline(aes(yintercept = 0.2))
theme(axis.text.x=element_text(angle=60, hjust=1))

If the file  does not exist (e.g. earthquakes), attempting to read this file will generate an error.
try(gplot() + geom_timeline_label(data=earthquakes, aes(Date, y=0.2, xend=xend, yend=yend, size=Mag, colour=Deaths,
label=Country), alpha=0.3) +
annotate("text", x=earthquakes_data_final$Date, y=0.42, label=earthquakes_data_final$Country, vjust=0, hjust=-0, size=2, angle=45) +
ylim(c(0,1)) + geom_hline(aes(yintercept = 0.2))
theme(axis.text.x=element_text(angle=60, hjust=1)))
}
