% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline class}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimeline
}
\value{
The ggproto function returns an object of class GeomTimeline for input as the default geom for
the user facing geom_timeline() function.
}
\description{
A new class corresponding to the geom_timeline geom.
}
\keyword{datasets}
