% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqks_functions.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel class}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimelineLabel
}
\value{
The ggproto function returns an object of class GeomTimelineLabel for input as the default geom for
the user facing geom_timeline_label() function.
}
\description{
A new class corresponding to the geom_timeline_label geom.
}
\note{
Example: ggproto("GeomTimelineLabel", ggplot2::Geom, required_aes = c("x", "y", "xend", "yend"),
draw_panel = function(data, panel_params, coord, ...) {
point1 = transform(data)
point2 = transform(data, x=xend, y=yend) }
)
}
\keyword{datasets}
